move orml config
