+++ runtime/selendra/src/config/orml.rs
@@ -1,8 +1,11 @@
 use crate::{
-	parameter_types, weights, AuctionManager, Balance, CurrencyId, Event, Incentives, Runtime,
+	parameter_types, weights, AuctionManager, Balance, CurrencyId, Event, Incentives, Runtime, Amount, ExistentialDeposits,
+	AccountId, MaxReserves, get_all_module_accounts, TreasuryAccount, MaxLocks, ReserveIdentifier, Moment, Timestamp,
+	Price, OperatorMembershipSelendra
 };
 use module_support::PoolId;
 use primitives::AuctionId;
+use frame_support::traits::Contains;
 
 impl orml_auction::Config for Runtime {
 	type Event = Event;
@@ -32,4 +35,49 @@ impl orml_nft::Config for Runtime {
 	type TokenData = module_nft::TokenData<Balance>;
 	type MaxClassMetadata = MaxTokenMetadata;
 	type MaxTokenMetadata = MaxTokenMetadata;
+}
+
+pub struct DustRemovalWhitelist;
+impl Contains<AccountId> for DustRemovalWhitelist {
+	fn contains(a: &AccountId) -> bool {
+		get_all_module_accounts().contains(a)
+	}
+}
+
+impl orml_tokens::Config for Runtime {
+	type Event = Event;
+	type Balance = Balance;
+	type Amount = Amount;
+	type CurrencyId = CurrencyId;
+	type ExistentialDeposits = ExistentialDeposits;
+	type OnDust = orml_tokens::TransferDust<Runtime, TreasuryAccount>;
+	type MaxLocks = MaxLocks;
+	type MaxReserves = MaxReserves;
+	type ReserveIdentifier = ReserveIdentifier;
+	type DustRemovalWhitelist = DustRemovalWhitelist;
+	type OnNewTokenAccount = ();
+	type OnKilledTokenAccount = ();
+	type WeightInfo = weights::orml_tokens::WeightInfo<Runtime>;
+}
+
+parameter_types! {
+	pub const MinimumCount: u32 = 5;
+	pub const ExpiresIn: Moment = 1000 * 60 * 60; // 1 hours
+	pub const MaxHasDispatchedSize: u32 = 40;
+	pub RootOperatorAccountId: AccountId = AccountId::from([0xffu8; 32]);
+}
+
+type SelendraDataProvider = orml_oracle::Instance1;
+impl orml_oracle::Config<SelendraDataProvider> for Runtime {
+	type Event = Event;
+	type OnNewData = ();
+	type CombineData =
+		orml_oracle::DefaultCombineData<Runtime, MinimumCount, ExpiresIn, SelendraDataProvider>;
+	type Time = Timestamp;
+	type OracleKey = CurrencyId;
+	type OracleValue = Price;
+	type RootOperatorAccountId = RootOperatorAccountId;
+	type Members = OperatorMembershipSelendra;
+	type MaxHasDispatchedSize = MaxHasDispatchedSize;
+	type WeightInfo = weights::orml_oracle::WeightInfo<Runtime>;
 }
